+++ src/Fable.Compiler/CHANGELOG.md
@@ -1,11 +1,25 @@
+---
+last_commit_released: 64c4dfdc595ed6720fcd053bee1f2d2930e4ead1
+include:
+  - ../fable-library-beam/
+  - ../fable-library-dart/
+  - ../fable-library-php/
+  - ../fable-library-php/
+  - ../fable-library-py/
+  - ../fable-library-rust/
+  - ../fable-library-ts/
+  - ../Fable.Compiler/
+  - ../Fable.Transforms/
+  - ../Fable.AST/
+  - ../../lib/
+---
+
 # Changelog
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-## Unreleased
 
 ## 5.0.0 - 2026-04-21