+++ src/fable-library-ts/CHANGELOG.md
@@ -1,11 +1,19 @@
+---
+last_commit_released: 514fc3601aba2d9bb608a520cb16d2819fa2eb28
+updaters:
+  - package.json:
+      file: package.json
+  - regex:
+      file: ./../Fable.Transforms/Global/Compiler.fs
+      pattern: (?<=let JS_LIBRARY_VERSION = ").*(?=")
+---
+
 # Changelog
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-## Unreleased
 
 ## 2.0.0 - 2026-04-21