+++ src/fable-standalone/CHANGELOG.md
@@ -1,11 +1,17 @@
+---
+last_commit_released: 64c4dfdc595ed6720fcd053bee1f2d2930e4ead1
+include:
+  - ../Fable.Transforms/
+  - ../fcs-fable/
+  - ../Fable.AST/
+---
+
 # Changelog
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-## Unreleased
 
 ## 2.0.0 - 2026-04-21